fix customer_id to BigInteger

+++ app/database/models/tables/customer.py
@@ -18,7 +18,7 @@ from database.engine import Base
 class Customer(Base):
     __tablename__ = 'customers'
 
-    customer_id: int = Column(Integer, primary_key=True, autoincrement=True)
+    customer_id: int = Column(BigInteger, primary_key=True, autoincrement=True)
     balance: float = Column(
         Float(precision=7, decimal_return_scale=2),
         nullable=False,
